refactor(agent): move filewriter under agent/tools
Move-only: the safe file writer, path validator, backup manager and streaming
chunk manager exist for one consumer - agent/tools' Write tool. Nesting it
under its owner says so.
