+++ internal/agent/tools/filewriter/chunks.go
