+++ internal/agent/tools/filewriter/contracts.go
