+++ internal/agent/tools/filewriter/validator.go
