+++ internal/agent/tools/filewriter/validator_test.go
