+++ internal/agent/tools/filewriter/writer.go
