+++ internal/agent/tools/write.go
@@ -11,7 +11,7 @@ import (
 	config "github.com/inference-gateway/cli/config"
 	agentdomain "github.com/inference-gateway/cli/internal/agent/domain"
 	agentinfra "github.com/inference-gateway/cli/internal/agent/infrastructure"
-	filewriter "github.com/inference-gateway/cli/internal/services/filewriter"
+	filewriter "github.com/inference-gateway/cli/internal/agent/tools/filewriter"
 )
 
 const (
